Implemented heuristically evaluation function
Evaluates a current board layout as described in the task paper.

+++ Task4/Spielbrett.h
@@ -23,7 +23,7 @@ class Spielbrett
 		Feld& operator () (size_t, size_t);       			// Zugriff auf Einträge der Matrix
 		
 		void setzeStein(int , Feld);
-		int heuristischeBewertung(Feld); 					//Bewertet die aktuelle Verteilung der Steine fuer eine gegebene Farbe
+		double heuristischeBewertung(Feld); 					//Bewertet die aktuelle Verteilung der Steine fuer eine gegebene Farbe
 
 		static void BrettFehler (const std::string& str);   // Fehlermeldung ausgeben
 };
